# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "juny-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "description": "OAuth login endpoint.\n\nVerify OAuth token, create/update user, and issue JWT tokens.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Refresh Token",
        "description": "Refresh access token using refresh token.",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/relations": {
      "post": {
        "tags": [
          "relations"
        ],
        "summary": "Create Care Relation",
        "description": "Create a new care relation between host and caregiver.\n\nOnly the host or an existing caregiver of the host may create new relations.",
        "operationId": "create_care_relation_api_v1_relations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareRelationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareRelationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "relations"
        ],
        "summary": "List Care Relations",
        "description": "List care relations filtered by host or caregiver.\n\nUsers can only list relations they participate in.",
        "operationId": "list_care_relations_api_v1_relations_get",
        "parameters": [
          {
            "name": "host_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host Id"
            }
          },
          {
            "name": "caregiver_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Caregiver Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareRelationResponse"
                  },
                  "title": "Response List Care Relations Api V1 Relations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wellness": {
      "post": {
        "tags": [
          "wellness"
        ],
        "summary": "Create Wellness Log",
        "description": "Create a wellness log entry for a host.",
        "operationId": "create_wellness_log_api_v1_wellness_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WellnessLogCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellnessLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "wellness"
        ],
        "summary": "List Wellness Logs",
        "description": "List wellness logs for a host (paginated, newest first).",
        "operationId": "list_wellness_logs_api_v1_wellness_get",
        "parameters": [
          {
            "name": "host_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Host Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WellnessLogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/medications": {
      "post": {
        "tags": [
          "medications"
        ],
        "summary": "Create Medication",
        "description": "Create a medication schedule entry for a host.",
        "operationId": "create_medication_api_v1_medications_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "medications"
        ],
        "summary": "List Medications",
        "description": "List medications for a host (paginated, newest first).",
        "operationId": "list_medications_api_v1_medications_get",
        "parameters": [
          {
            "name": "host_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Host Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MedicationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CareRelationCreate": {
        "properties": {
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "caregiver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Caregiver Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Caregiver role",
            "examples": [
              "concierge",
              "care_worker",
              "organization"
            ]
          }
        },
        "type": "object",
        "required": [
          "host_id",
          "caregiver_id",
          "role"
        ],
        "title": "CareRelationCreate",
        "description": "Request body for creating a care relation."
      },
      "CareRelationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "caregiver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Caregiver Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "host_id",
          "caregiver_id",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "CareRelationResponse",
        "description": "Response model for a care relation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicationCreate": {
        "properties": {
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "pill_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Pill Name"
          },
          "schedule_time": {
            "type": "string",
            "format": "date-time",
            "title": "Schedule Time"
          }
        },
        "type": "object",
        "required": [
          "host_id",
          "pill_name",
          "schedule_time"
        ],
        "title": "MedicationCreate",
        "description": "Request body for creating a medication schedule."
      },
      "MedicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "pill_name": {
            "type": "string",
            "title": "Pill Name"
          },
          "schedule_time": {
            "type": "string",
            "format": "date-time",
            "title": "Schedule Time"
          },
          "is_taken": {
            "type": "boolean",
            "title": "Is Taken"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "host_id",
          "pill_name",
          "schedule_time",
          "is_taken",
          "created_at"
        ],
        "title": "MedicationResponse",
        "description": "Response model for a medication entry."
      },
...
      "OAuthLoginRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "google",
              "github",
              "facebook"
            ],
            "title": "Provider"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "access_token"
        ],
        "title": "OAuthLoginRequest",
        "description": "OAuth login request."
      },
      "PaginatedResponse_MedicationResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MedicationResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PaginatedResponse[MedicationResponse]"
      },
      "PaginatedResponse_WellnessLogResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WellnessLogResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PaginatedResponse[WellnessLogResponse]"
      },
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Items per page"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "Total number of pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "description": "Whether there is a next page"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev",
            "description": "Whether there is a previous page"
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "total_pages",
          "has_next",
          "has_prev"
        ],
        "title": "PaginationMeta",
        "description": "Pagination metadata."
      },
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "description": "Refresh token request."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse",
        "description": "Token response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WellnessLogCreate": {
        "properties": {
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "status": {
            "$ref": "#/components/schemas/WellnessStatus"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "host_id",
          "status",
          "summary"
        ],
        "title": "WellnessLogCreate",
        "description": "Request body for creating a wellness log."
      },
      "WellnessLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "host_id",
          "status",
          "summary",
          "details",
          "created_at"
        ],
        "title": "WellnessLogResponse",
        "description": "Response model for a wellness log."
      },
      "WellnessStatus": {
        "type": "string",
        "enum": [
          "normal",
          "warning",
          "emergency"
        ],
        "title": "WellnessStatus",
        "description": "Wellness log severity levels."
      }
    }
  }
}